requirements.txt
torch>=1.0
numpy>=1.16
Pillow>=5.4
tensorboardX>=1.6
